refactor: split tui.rs into mod, render, and tests submodules
Split the 2137-line src/tui.rs into three focused files following
the same pattern used for the conductor refactor:
- src/tui/mod.rs: types, App struct + impl, action constants
- src/tui/render.rs: all draw/UI helper free functions (pub(super) for tested helpers)
- src/tui/tests.rs: all 75 tests extracted from inline mod block
